@@ -15,9 +15,8 @@
     <description>Demo project for Spring Boot</description>
 
     <properties>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
         <repackage.classifier/>
-        <spring-native.version>0.11.2</spring-native.version>
     </properties>
 
     <dependencies>
