@@ -1,25 +1,28 @@
 package io.axoniq.demo.bikerental.bikerental;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.thoughtworks.xstream.XStream;
 import io.axoniq.demo.bikerental.bikerental.history.BikeHistory;
 import io.axoniq.demo.bikerental.bikerental.query.BikeStatus;
-import io.grpc.EquivalentAddressGroup;
 import org.axonframework.common.jdbc.ConnectionProvider;
 import org.axonframework.config.Configuration;
-import org.axonframework.config.Configurer;
-import org.axonframework.eventhandling.tokenstore.ConfigToken;
+import org.axonframework.config.EventProcessingConfigurer;
 import org.axonframework.eventhandling.tokenstore.jdbc.GenericTokenTableFactory;
 import org.axonframework.eventhandling.tokenstore.jdbc.JdbcTokenStore;
+import org.springframework.aot.hint.MemberCategory;
+import org.springframework.aot.hint.RuntimeHints;
+import org.springframework.aot.hint.RuntimeHintsRegistrar;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Bean;
-import org.springframework.nativex.hint.NativeHint;
-import org.springframework.nativex.hint.TypeHint;
+import org.springframework.context.annotation.ImportRuntimeHints;
 
-@TypeHint(types = BikeStatus.class)
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+
+@ImportRuntimeHints(BikerentalApplication.HintsRegistrar.class)
 @EntityScan(basePackageClasses = {BikeHistory.class, BikeStatus.class})
 @SpringBootApplication
 public class BikerentalApplication {
@@ -30,7 +33,42 @@ public class BikerentalApplication {
 
     @Autowired
     public void configure(ObjectMapper objectMapper) {
-        objectMapper.activateDefaultTyping(objectMapper.getPolymorphicTypeValidator());;
+        objectMapper.activateDefaultTyping(objectMapper.getPolymorphicTypeValidator(),
+                                           ObjectMapper.DefaultTyping.NON_CONCRETE_AND_ARRAYS);
+    }
+
+    @Autowired
+    public void eventProcessingModule(EventProcessingConfigurer configurer) {
+        configurer.usingPooledStreamingEventProcessors()
+                  .registerPooledStreamingEventProcessorConfiguration((c, processor) ->
+                                                                              processor.workerExecutor(workerExecutor()));
+    }
+//
+//    @Bean
+//    public EventProcessorControlService processorProvider(AxonServerConnectionManager axonServerConnectionManager,
+//                                                          EventProcessingConfiguration eventProcessingConfiguration,
+//                                                          Configuration axonConfig,
+//                                                          AxonServerConfiguration configuration) {
+//        var eventProcessorControlService = new EventProcessorControlService(axonServerConnectionManager, eventProcessingConfiguration, configuration.getContext()) {
+//            @Override
+//            public Supplier<EventProcessorInfo> infoSupplier(EventProcessor processor) {
+//                Supplier<EventProcessorInfo> parentSupplier = super.infoSupplier(processor);
+//                return () -> {
+//                    EventProcessorInfo parentInfo = parentSupplier.get();
+////                    logger.info("Retrieved information: {}", parentInfo);
+//                    return parentInfo;
+//                };
+//            }
+//        };
+//        logger.info("Control Service registered.");
+//        axonConfig.onStart(Phase.INSTRUCTION_COMPONENTS + 1, eventProcessorControlService::start);
+//        return eventProcessorControlService;
+//    }
+
+    @Qualifier("workerExecutor")
+    @Bean(destroyMethod = "shutdown")
+    public ScheduledExecutorService workerExecutor() {
+        return Executors.newScheduledThreadPool(2);
     }
 
     @Bean
@@ -41,6 +79,16 @@ public class BikerentalApplication {
                                        .build();
         configuration.onStart(Integer.MIN_VALUE, () -> tokenStore.createSchema(GenericTokenTableFactory.INSTANCE));
         return tokenStore;
+    }
+
+    public static class HintsRegistrar implements RuntimeHintsRegistrar {
+        @Override
+        public void registerHints(RuntimeHints hints, ClassLoader classLoader) {
+            hints.reflection().registerType(BikeStatus.class,
+                                            MemberCategory.DECLARED_FIELDS,
+                                            MemberCategory.INTROSPECT_DECLARED_METHODS,
+                                            MemberCategory.INTROSPECT_DECLARED_CONSTRUCTORS);
+        }
     }
 
 }
